Remove redundant console logging in test fixture. Convert to ts.
